+++ HISTORY.md
@@ -4,6 +4,8 @@
 * Add additional protected methods to JsonCodec to allow subclasses to
   further customize the encoding used.
 * Allow objects to have numeric properties.
+* Ensure objects with numeric properties are consistently encoded using
+  the '{...}' JSON syntax.
 
 ## 2.1.0 (2023-10-02)
 * Allow ::jsonClassHintFor() to return a class-string suffixed with